requirements.txt
django-flexible-content
django-taggit
django-nose
coverage
pep8
pylint
